requirements.txt

# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
tweepy==4.14.0
youtube-transcript-api==0.6.1
pytube==15.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2